requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.0
